requirements.txt
# Core ML and Data Processing
tensorflow>=2.12.0
keras>=2.12.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
Pillow>=8.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Web Framework and API
flask>=2.0.0
streamlit>=1.25.0
flask-restful>=0.3.9
gunicorn>=20.1.0

# Image Processing and Medical Imaging
SimpleITK>=2.1.0
pydicom>=2.3.0
nibabel>=3.2.0

# Model Interpretability
grad-cam>=1.4.0

# Development and Testing
pytest>=6.2.0
jupyter>=1.0.0
tqdm>=4.62.0

# Visualization
plotly>=5.0.0